cluster_load_status gets load avg from pbs
Updated cluster_load_status to handle getting load average information from
pbs and to remove the need to explicitly specify all hosts in the one file
(this is now contained in a nodes file which is read in)
Removed the old explicit node information: now everything is in the nodes
file
Since we are now getting the load average information from pbs, we don't
need the ssh getload routine: removed
Now reading and processing the input xml file.  Added help information and
command line options for this.  Also cleaned up the command line option
parsing code; too many if's were being used.
When testing the code the x and y position of the node in the matrix is now
used to determine the "load average"